move read functions to addresses pkg
