+++ packages/addresses/hardhat.config.ts
@@ -1,0 +1,5 @@
+import { HardhatUserConfig } from "hardhat/types";
+
+const config: HardhatUserConfig = {};
+
+export default config;
