+++ packages/protocol-contracts/scripts/deploy-immutable-create2-factory.ts
@@ -1,4 +1,5 @@
-import { isNetworkName, saveAddress } from "@zetachain/addresses";
+import { isNetworkName } from "@zetachain/addresses";
+import { saveAddress } from "@zetachain/addresses-tools";
 import { network } from "hardhat";
 
 import { deployImmutableCreate2Factory as deploy } from "../lib/contracts.helpers";
